Arya

+++ features/rooms/service/room_service.go
@@ -2,31 +2,50 @@ package service
 
 import (
 	"errors"
+	"project/features/instruments"
 	"project/features/rooms"
+	"project/helper"
+
+	"github.com/golang-jwt/jwt/v5"
 )
 
 type RoomService struct {
 	d rooms.RoomDataInterface
+	j helper.JWTInterface
 }
 
-func NewRoomService(data rooms.RoomDataInterface) rooms.RoomService {
+func NewRoomService(data rooms.RoomDataInterface, jwt helper.JWTInterface) rooms.RoomService {
 	return &RoomService{
 		d: data,
+		j: jwt,
 	}
 }
 
-func (rs *RoomService) AddRoom(newRoom rooms.Rooms) (*rooms.Rooms, error) {
-	res, err := rs.d.AddRoom(newRoom)
-	if err != nil {
-		return nil, errors.New("Cannot Add Room")
+func (rs *RoomService) AddRoom(newRoom rooms.Rooms, token *jwt.Token) (*rooms.Rooms, error) {
+	if _, role := rs.j.ExtractToken(token); role != "admin" {
+		return nil, errors.New("Unauthorized user")
 	}
+
+	var currentRoom rooms.Rooms
+
+	currentRoom.Price = newRoom.Price
+	currentRoom.RoomID = newRoom.RoomID
+
+	res, err := rs.d.AddRoom(currentRoom)
+	if err != nil {
+		return nil, err
+	}
+
 	return res, nil
 }
 
-func (rs *RoomService) DeleteRoom(roomID string) (string, error) {
+func (rs *RoomService) DeleteRoom(roomID string, token *jwt.Token) (any, error) {
+	if _, role := rs.j.ExtractToken(token); role != "admin" {
+		return nil, errors.New("Unauthorized user")
+	}
 	err := rs.d.DeleteRoom(roomID)
 	if err != nil {
-		return "", errors.New("Cannot delete room")
+		return nil, errors.New("Cannot delete room")
 	}
 	return roomID, nil
 }
@@ -51,6 +70,18 @@ func (rs *RoomService) UpdateRoom(roomID string, updatedRoom rooms.Rooms) (*room
 	res, err := rs.d.UpdateRoom(roomID, updatedRoom)
 	if err != nil {
 		return nil, errors.New("Cannot update room")
+	}
+	return res, nil
+}
+
+func (rs *RoomService) AddRoomInstrument(roomID string, instrumentData instruments.RoomInstrument, token *jwt.Token) (any, error) {
+	_, role := rs.j.ExtractToken(token)
+	if role != "admin" {
+		return nil, errors.New("Unautorized user")
+	}
+	res, err := rs.d.AddRoomInstrument(roomID, instrumentData)
+	if err != nil {
+		return nil, err
 	}
 	return res, nil
 }
